+++ game.c
@@ -31,7 +31,7 @@ enum ships {
 
 typedef struct {
 	size_t size;
-	uint16_t initial_row, initial_column;
+	uint8_t initial_row, initial_column;
 	bool direction; // 1: vertical, 0: horizontal
 	enum ships type;
 } ship_st;
@@ -197,10 +197,10 @@ bool valid_position(size_t board_size, uint8_t *board, ship_st ship)
 /**
  * Converte coordenadas para índices da matriz do tabuleiro
  */
-void convert_coords_to_index(char tmp_col, uint16_t tmp_row, uint16_t *col, uint16_t *row)
+void convert_coords_to_index(char tmp_col, uint16_t tmp_row, uint8_t *col, uint8_t *row)
 {
-	*col = (uint16_t) toupper(tmp_col) - 65;
-	*row = tmp_row - 1;
+	*col = (uint8_t) (toupper(tmp_col) - 65);
+	*row = (uint8_t) (tmp_row - 1);
 }
 
 /**
@@ -290,7 +290,7 @@ game_st set_default_values()
  */
 void shot_try(size_t board_size, player_st *player, player_st *enemy, bool human)
 {
-	uint16_t row, col;
+	uint8_t row, col;
 
 	// Le e valida coordenadas
 	do {
