+++ game.h
@@ -2,7 +2,7 @@
 #define __GAME_H__
 
 void game_new();
-void convert_coords_to_index(char, uint16_t, uint16_t*, uint16_t*);
+void convert_coords_to_index(char, uint16_t, uint8_t*, uint8_t*);
 
 #endif
